Added options to specify number of threads OpenMP uses
CRAN doesn't allow more than two threads in their test
farm; users can be completely uneducated about setting
environment variable as well, especially when they run
multiple models through snow. For these reasons now the
number of threads is defaulted to two and it's the users'
job to specify how many threads to use.

But as a result of this, now the package doesn't
automatically detect how many cores there are any more.
The previous mechanism for detection was to let OpenMP
figure out with their magic, but this magic isn't exposed
as API AFAIK. And now without this, it seems very hard
to detect it in a portable way.